PR feedback: move DocAttachment to attachment.go
